ai_assistant/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.0.1
scikit-learn==1.3.2
motor==3.3.1
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.21
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
google-generativeai==0.3.2 